Cyclistic: Data Analyst Case Study
This case study focuses on a fictional company called Cyclistic, which aims to maximize its annual membership by converting casual riders into membership customers. The study includes a PDF file with internal links directing to the data sets, as well as a final presentation containing visualizations, insights, and recommendations for the marketing team.